pyproject.toml
[project]
name = "koukoutu"
description = ""
version = "1.0.2"
license = {file = "LICENSE"} 
# classifiers = [
#     # For OS-independent nodes (works on all operating systems)
#     "Operating System :: OS Independent",
#
#     # OR for OS-specific nodes, specify the supported systems:
#     "Operating System :: Microsoft :: Windows",  # Windows specific
#     "Operating System :: POSIX :: Linux",  # Linux specific
#     "Operating System :: MacOS",  # macOS specific
#
#     # GPU Accelerator support. Pick the ones that are supported by your extension.
#     "Environment :: GPU :: NVIDIA CUDA",    # NVIDIA CUDA support
#     "Environment :: GPU :: AMD ROCm",       # AMD ROCm support
#     "Environment :: GPU :: Intel Arc",      # Intel Arc support
#     "Environment :: NPU :: Huawei Ascend",  # Huawei Ascend support
#     "Environment :: GPU :: Apple Metal",    # Apple Metal support
# ]

dependencies = ["requests", "Pillow", "numpy", "torch"]

[project.urls]
Repository = "https://github.com/zhenzi0322/comfyui-koukoutu"
#  Used by Comfy Registry https://registry.comfy.org
Documentation = "https://github.com/zhenzi0322/comfyui-koukoutu/wiki"
"Bug Tracker" = "https://github.com/zhenzi0322/comfyui-koukoutu/issues"

[tool.comfy]
PublisherId = "zhenzi0322"
DisplayName = "comfyui-koukoutu"
Icon = ""
includes = [] 
# "requires-comfyui" = ">=1.0.0"  # ComfyUI version compatibility

